# client/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Loom.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/exchanges/{exchangeGuid}": {
      "get": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateExchangeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/exchanges": {
      "post": {
        "tags": [
          "Exchange"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateExchangeRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/coordinator-message": {
      "put": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateCoordinatorMessageRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/access-link": {
      "post": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/access-link/regenerate": {
      "post": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLinkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/partner-courses": {
      "get": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Semester",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ExchangeSemester"
                }
              ]
            }
          },
          {
            "name": "Level",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StudyProgramLevel"
                }
              ]
            }
          },
          {
            "name": "IncludeDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDir",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SafePage",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SafePageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponsePage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exchange"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PartnerCourseRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerCourseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exchanges/{exchangeGuid}/recognition": {
      "get": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/recognition/start": {
      "post": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/recognition/grades": {
      "put": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/SaveGradesRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/recognition/status": {
      "patch": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateRecognitionStatusRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/recognition/message": {
      "patch": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateRecognitionMessageRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/exchanges/{exchangeGuid}/recognition/versions": {
      "get": {
        "tags": [
          "Recognition"
        ],
        "parameters": [
          {
            "name": "exchangeGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentVersionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessLinkResponse": {
        "required": [
          "createdAt",
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CourseGradesRequest": {
        "required": [
          "partnerCourseId"
        ],
        "type": "object",
        "properties": {
          "partnerCourseId": {
            "type": "integer",
            "format": "int32"
          },
          "enrollmentStatus": {
            "type": "string",
            "nullable": true
          },
          "originalGrade": {
            "type": "string",
            "nullable": true
          },
          "ectsGrade": {
            "type": "string",
            "nullable": true
          },
          "hrGrade": {
            "type": "string",
            "nullable": true
          },
          "examDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateExchangeRequest": {
        "required": [
          "academicYear",
          "homeProfileId",
          "partnerInstitutionId",
          "semesterType",
          "studySemesters"
        ],
        "type": "object",
        "properties": {
          "homeProfileId": {
            "type": "integer",
            "format": "int32"
          },
          "partnerInstitutionId": {
            "type": "integer",
            "format": "int32"
          },
          "academicYear": {
            "type": "string"
          },
          "semesterType": {
            "type": "string"
          },
          "studySemesters": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetStudentId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DocumentStatus": {
        "enum": [
          "Draft",
          "Approved"
        ],
        "type": "string"
      },
      "DocumentVersionResponse": {
        "required": [
          "amendmentLabel",
          "changes",
          "createdAt",
          "createdByName",
          "entryCount",
          "id",
          "kind",
          "versionNo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/VersionKind"
              }
            ]
          },
          "versionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "amendmentLabel": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByName": {
            "type": "string",
            "nullable": true
          },
          "entryCount": {
            "type": "integer",
            "format": "int32"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentChange"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExchangeResponse": {
        "required": [
          "academicYear",
          "coordinatorId",
          "coordinatorMessage",
          "coordinatorName",
          "createdAt",
          "ewpLink",
          "guid",
          "homeInstitutionName",
          "homeProfile",
          "homeProgramName",
          "id",
          "mentor",
          "partnerInstitutionId",
          "partnerInstitutionName",
          "semesterType",
          "studentId",
          "studentIsPlaceholder",
          "studentJmbag",
          "studentName",
          "studySemesters",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "guid": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "studentName": {
            "type": "string"
          },
          "studentJmbag": {
            "type": "string",
            "nullable": true
          },
          "homeInstitutionName": {
            "type": "string"
          },
          "homeProgramName": {
            "type": "string"
          },
          "homeProfile": {
            "allOf": [
              {
                "$ref": "#/components/schemas/HomeProfileResponse"
              }
            ]
          },
          "partnerInstitutionId": {
            "type": "integer",
            "format": "int32"
          },
          "partnerInstitutionName": {
            "type": "string"
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coordinatorName": {
            "type": "string",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          },
          "academicYear": {
            "type": "string"
          },
          "semesterType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeSemester"
              }
            ]
          },
          "studySemesters": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "coordinatorMessage": {
            "type": "string",
            "nullable": true
          },
          "ewpLink": {
            "type": "string",
            "nullable": true
          },
          "studentIsPlaceholder": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ExchangeSemester": {
        "enum": [
          "Winter",
          "Summer",
          "Both"
        ],
        "type": "string"
      },
...
      "HomeProfileResponse": {
        "required": [
          "id",
          "name",
          "nameEn"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "nameEn": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PartnerCourseRequest": {
        "required": [
          "code",
          "ects",
          "level",
          "name",
          "semester"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "ects": {
            "type": "number",
            "format": "double"
          },
          "semester": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "lecturesH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "auditoryH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "labH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseResponse": {
        "required": [
          "auditoryH",
          "code",
          "ects",
          "id",
          "isDeleted",
          "labH",
          "lecturesH",
          "level",
          "name",
          "nameHr",
          "semester",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "nameHr": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "ects": {
            "type": "number",
            "format": "double"
          },
          "lecturesH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "auditoryH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "labH": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "semester": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExchangeSemester"
              }
            ]
          },
          "level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StudyProgramLevel"
              }
            ]
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "PartnerCourseResponsePage": {
        "required": [
          "hasDeleted",
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartnerCourseResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "RecognitionResponse": {
        "required": [
          "agreed",
          "agreedVersionNo",
          "approvedVersionCount",
          "canStart",
          "exchangeId",
          "isStarted",
          "lastModifiedAt",
          "lastModifiedByName",
          "message",
          "signedAt",
          "signedByName",
          "startedAt",
          "startedByName",
          "status"
        ],
        "type": "object",
        "properties": {
          "exchangeId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DocumentStatus"
              }
            ]
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "isStarted": {
            "type": "boolean"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startedByName": {
            "type": "string",
            "nullable": true
          },
          "canStart": {
            "type": "boolean"
          },
          "agreedVersionNo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "agreed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgreedEntryResponse"
            }
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModifiedByName": {
            "type": "string",
            "nullable": true
          },
          "signedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "signedByName": {
            "type": "string",
            "nullable": true
          },
          "approvedVersionCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SaveGradesRequest": {
        "required": [
          "entries"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseGradesRequest"
            }
          }
        },
        "additionalProperties": false
      },
...
      "StudyProgramLevel": {
        "enum": [
          "Undergraduate",
          "Graduate",
          "Postgraduate"
        ],
        "type": "string"
      },
      "UpdateCoordinatorMessageRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateExchangeRequest": {
        "required": [
          "academicYear",
          "semesterType",
          "studySemesters"
        ],
        "type": "object",
        "properties": {
          "academicYear": {
            "type": "string"
          },
          "semesterType": {
            "type": "string"
          },
          "studySemesters": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "coordinatorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "mentor": {
            "type": "string",
            "nullable": true
          },
          "ewpLink": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRecognitionMessageRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateRecognitionStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "VersionKind": {
        "enum": [
          "Approved",
          "Backup"
        ],
        "type": "string"
      }
    }
  },
...
}